Improved test structure
