+++ SchoolManagerTests/ClassManagerTests.cs
@@ -3,17 +3,17 @@ using SchoolManagerModel;
 using SchoolManagerModel.Persistence;
 using SchoolManagerModel.UserModel;
 
-namespace SchoolManager.Tests
+namespace SchoolManagerTests
 {
     [TestClass]
     public class ClassManagerTests
     {
-        private Mock<IAsyncClassDataHandler> _mockDataHandler = new Mock<IAsyncClassDataHandler>();
-        private ClassManager _classManager;
+        private readonly Mock<IAsyncClassDataHandler> _mockDataHandler = new Mock<IAsyncClassDataHandler>();
+        private readonly ClassManager _classManager;
 
-        private Class _testClass = new Class { Id = 1, Name = "1/A" };
-        private User _testStudent = new User("username", "password", "email", "firstName", "lastName");
-        private Subject _testSubject;
+        private readonly Class _testClass = new Class { Id = 1, Name = "1/A" };
+        private readonly User _testStudent = new User("username", "password", "email", "firstName", "lastName");
+        private readonly Subject _testSubject;
 
         public ClassManagerTests()
         {
@@ -29,23 +29,15 @@ namespace SchoolManager.Tests
                 Teacher = new Teacher { User = new User("teacher", "pass", "email", "firstName", "lastName") }
             };
         }
-
-        public void Setup()
-        {
-
-        }
-
+        
         [TestMethod]
         public async Task GetClassesAsync_ShouldReturnClasses()
         {
-            // Arrange
             var classes = new List<Class> { _testClass };
             _mockDataHandler.Setup(dh => dh.GetClassesAsync()).ReturnsAsync(classes);
-
-            // Act
+            
             var result = await _classManager.GetClassesAsync();
-
-            // Assert
+            
             Assert.AreEqual(1, result.Count);
             Assert.AreEqual(_testClass, result[0]);
             _mockDataHandler.Verify(dh => dh.GetClassesAsync(), Times.Once);
@@ -54,13 +46,8 @@ namespace SchoolManager.Tests
         [TestMethod]
         public async Task AddClassAsync_ShouldAddClass_WhenClassDoesNotExist()
         {
-            // Arrange
             _mockDataHandler.Setup(dh => dh.ClassExistsAsync(_testClass)).ReturnsAsync(false);
-
-            // Act
             await _classManager.AddClassAsync(_testClass);
-
-            // Assert
             _mockDataHandler.Verify(dh => dh.AddClassAsync(_testClass), Times.Once);
         }
 
@@ -68,26 +55,17 @@ namespace SchoolManager.Tests
         [ExpectedException(typeof(ClassExistsException))]
         public async Task AddClassAsync_ShouldThrowException_WhenClassExists()
         {
-            // Arrange
             _mockDataHandler.Setup(dh => dh.ClassExistsAsync(_testClass)).ReturnsAsync(true);
-
-            // Act
             await _classManager.AddClassAsync(_testClass);
-
-            // Assert - Expect exception
         }
 
         [TestMethod]
         public async Task GetClassStudentsAsync_ShouldReturnClassStudents()
         {
-            // Arrange
             var students = new List<User> { _testStudent };
             _mockDataHandler.Setup(dh => dh.GetClassStudentsAsync(_testClass)).ReturnsAsync(students);
-
-            // Act
+            
             var result = await _classManager.GetClassStudentsAsync(_testClass);
-
-            // Assert
             Assert.AreEqual(1, result.Count);
             Assert.AreEqual(_testStudent, result[0]);
             _mockDataHandler.Verify(dh => dh.GetClassStudentsAsync(_testClass), Times.Once);
@@ -96,14 +74,10 @@ namespace SchoolManager.Tests
         [TestMethod]
         public async Task GetClassSubjectsAsync_ShouldReturnClassSubjects()
         {
-            // Arrange
             var subjects = new List<Subject> { _testSubject };
             _mockDataHandler.Setup(dh => dh.GetClassSubjectsAsync(_testClass)).ReturnsAsync(subjects);
-
-            // Act
+            
             var result = await _classManager.GetClassSubjectsAsync(_testClass);
-
-            // Assert
             Assert.AreEqual(1, result.Count);
             Assert.AreEqual(_testSubject, result[0]);
             _mockDataHandler.Verify(dh => dh.GetClassSubjectsAsync(_testClass), Times.Once);
