+++ SchoolManagerTests/UserManagerTests/AssignRoleAsyncTests.cs
@@ -38,6 +38,32 @@ public partial class UserManagerTests
         Assert.AreEqual(Role.Teacher, assignedRole);
 
     }
+    
+    
+    [TestMethod]
+    public async Task GetRoleAsync_ShouldReturnStudent_WhenRoleIsNull()
+    {
+        _handler.Setup(dh => dh.GetRoleAsync(_testUser)).ReturnsAsync((Role?)null);
+
+        var result = await _userManager.GetRoleAsync(_testUser);
+        Assert.AreEqual(Role.Student, result);
+    }
+
+    [TestMethod]
+    public async Task AssignRoleAsync_ShouldThrowException_WhenUserNotFound()
+    {
+        _handler.Setup(dh => dh.UsernameExistsAsync(_testUser.Username)).ReturnsAsync(false);
+        await Assert.ThrowsExceptionAsync<Exception>(() => _userManager.AssignRoleAsync(_testUser, Role.Teacher));
+    }
+
+    [TestMethod]
+    public async Task AssignRoleAsync_ShouldAssignRole_WhenUserExists()
+    {
+        _handler.Setup(dh => dh.UsernameExistsAsync(_testUser.Username)).ReturnsAsync(true);
+
+        await _userManager.AssignRoleAsync(_testUser, Role.Teacher);
+        _handler.Verify(dh => dh.AssignRoleAsync(_testUser, Role.Teacher), Times.Once);
+    }
 
     private void SetupUsernameExistsAsync(bool returnValue)
     {
