+++ SchoolManagerTests/UserManagerTests/RegisterUserAsyncTests.cs
@@ -37,5 +37,23 @@ public partial class UserManagerTests
             .Returns(Task.FromResult(users.Contains(user)));
         Assert.IsTrue(await _userManager.UserExistsAsync(user));
     }
+    
+    
+    [TestMethod]
+    public async Task RegisterUserAsync_ShouldThrowException_WhenUserAlreadyExists()
+    {
+        _handler.Setup(dh => dh.UsernameExistsAsync(_testUser.Username)).ReturnsAsync(true);
+        await Assert.ThrowsExceptionAsync<Exception>(() => _userManager.RegisterUserAsync(_testUser));
+    }
+
+    [TestMethod]
+    public async Task RegisterUserAsync_ShouldAddUser_WhenUserIsNew()
+    {
+        _handler.Setup(dh => dh.UsernameExistsAsync(_testUser.Username)).ReturnsAsync(false);
+        _handler.Setup(dh => dh.EmailAlreadyRegisteredAsync(_testUser.Email)).ReturnsAsync(false);
+
+        await _userManager.RegisterUserAsync(_testUser);
+        _handler.Verify(dh => dh.AddUserAsync(It.Is<User>(u => u.Username == _testUser.Username && u.Password != "password")), Times.Once);
+    }
 
 }
